Fix frequency / LNB precision issue when storing and loading source profiles

+++ analyzer/source.c
@@ -1407,17 +1407,17 @@ suscan_source_config_to_object(const suscan_source_config_t *cfg)
     SU_CFGSAVE(value, interface);
 
   /* XXX: This is terrible. Either change this or define SUFREQ as uint64_t */
-  SU_CFGSAVE(float, freq);
-  SU_CFGSAVE(float, lnb_freq);
-  SU_CFGSAVE(float, bandwidth);
-  SU_CFGSAVE(bool,  iq_balance);
-  SU_CFGSAVE(bool,  dc_remove);
-  SU_CFGSAVE(float, ppm);
-  SU_CFGSAVE(tv,    start_time);
-  SU_CFGSAVE(bool,  loop);
-  SU_CFGSAVE(uint,  samp_rate);
-  SU_CFGSAVE(uint,  average);
-  SU_CFGSAVE(uint,  channel);
+  SU_CFGSAVE(double, freq);
+  SU_CFGSAVE(double, lnb_freq);
+  SU_CFGSAVE(float,  bandwidth);
+  SU_CFGSAVE(bool,   iq_balance);
+  SU_CFGSAVE(bool,   dc_remove);
+  SU_CFGSAVE(float,  ppm);
+  SU_CFGSAVE(tv,     start_time);
+  SU_CFGSAVE(bool,   loop);
+  SU_CFGSAVE(uint,   samp_rate);
+  SU_CFGSAVE(uint,   average);
+  SU_CFGSAVE(uint,   channel);
 
   /* Save SoapySDR kwargs */
   SU_TRYCATCH(obj = suscan_object_new(SUSCAN_OBJECT_TYPE_OBJECT), goto fail);
@@ -1525,16 +1525,16 @@ suscan_source_config_from_object(const suscan_object_t *object)
     }
   }
 
-  SU_CFGLOAD(float, freq, 0);
-  SU_CFGLOAD(float, lnb_freq, 0);
-  SU_CFGLOAD(float, bandwidth, 0);
-  SU_CFGLOAD(bool,  iq_balance, SU_FALSE);
-  SU_CFGLOAD(bool,  dc_remove, SU_FALSE);
-  SU_CFGLOAD(float, ppm, 0);
-  SU_CFGLOAD(tv,    start_time, &default_time);
-  SU_CFGLOAD(bool,  loop, SU_FALSE);
-  SU_CFGLOAD(uint,  samp_rate, 1.8e6);
-  SU_CFGLOAD(uint,  channel, 0);
+  SU_CFGLOAD(double, freq, 0);
+  SU_CFGLOAD(double, lnb_freq, 0);
+  SU_CFGLOAD(float,  bandwidth, 0);
+  SU_CFGLOAD(bool,   iq_balance, SU_FALSE);
+  SU_CFGLOAD(bool,   dc_remove, SU_FALSE);
+  SU_CFGLOAD(float,  ppm, 0);
+  SU_CFGLOAD(tv,     start_time, &default_time);
+  SU_CFGLOAD(bool,   loop, SU_FALSE);
+  SU_CFGLOAD(uint,   samp_rate, 1.8e6);
+  SU_CFGLOAD(uint,   channel, 0);
 
   SU_TRYCATCH(SU_CFGLOAD(uint, average, 1), goto fail);
